fix: refacto errors
