+++ packages/datasource-sql/test/connection/handle-errors.unit.test.ts
@@ -2,6 +2,7 @@ import { ConnectionError } from 'sequelize';
 
 import ConnectionOptions from '../../src/connection/connection-options';
 import handleErrors from '../../src/connection/handle-errors';
+import { SocksProxyServiceError } from '../../src/connection/services/errors';
 
 const optionsWithoutProxy = new ConnectionOptions({
   uri: 'postgres://user:password@localhost:5432/db',
@@ -36,7 +37,7 @@ describe('handleErrors', () => {
   describe('when the error is thrown by the proxy', () => {
     describe('when the proxy has an error with database uri', () => {
       it('should throw an error', () => {
-        const error = new Error('Socket closed');
+        const error = new SocksProxyServiceError(new Error('Socket closed'));
         error.stack = 'SocksClient.connect: Socks connection failed to proxy: Socket closed';
 
         expect(() => handleErrors(error, optionsWithProxy)).toThrow(
@@ -48,7 +49,7 @@ describe('handleErrors', () => {
 
     describe('when the proxy config is null', () => {
       it('should throw an error', () => {
-        const error = new Error('An error');
+        const error = new SocksProxyServiceError(new Error('An error'));
         error.stack = 'SocksClient.connect: Socks connection failed to proxy: Socket closed';
 
         expect(() => handleErrors(error, optionsWithProxy)).toThrow(
@@ -60,7 +61,7 @@ describe('handleErrors', () => {
 
     describe('when the proxy has an error', () => {
       it('should throw an error', () => {
-        const error = new Error();
+        const error = new SocksProxyServiceError(new Error());
         error.stack = 'SocksClient.connect: Socks connection failed to proxy: Connection refused';
 
         expect(() => handleErrors(error, optionsWithProxy)).toThrow(
